move itemEnter & itemGetEntryState into the items package
