+++ app/api/items/enter.go
@@ -1,4 +1,4 @@
-package contests
+package items
 
 import (
 	"net/http"
