+++ app/api/items/get_entry_state.go
@@ -1,4 +1,4 @@
-package contests
+package items
 
 import (
 	"net/http"
